Move target iterator logic into uptane module
Iterating targets really isn't dependent on the SotaUptaneClient and
future code in the package manager will need to be able to iterate
through targets to function. Moving it now will help remove a future
circular dependency.
