+++ src/libaktualizr/primary/sotauptaneclient.h
@@ -21,63 +21,9 @@
 #include "reportqueue.h"
 #include "storage/invstorage.h"
 #include "uptane/directorrepository.h"
-#include "uptane/fetcher.h"
-#include "uptane/imagesrepository.h"
 #include "uptane/ipsecondarydiscovery.h"
+#include "uptane/iterator.h"
 #include "uptane/secondaryinterface.h"
-#include "uptane/uptanerepository.h"
-
-class LazyTargetsList {
- public:
-  struct DelegatedTargetTreeNode {
-    Uptane::Role role{Uptane::Role::Targets()};
-    DelegatedTargetTreeNode *parent{nullptr};
-    std::vector<std::shared_ptr<DelegatedTargetTreeNode>>::size_type parent_idx{0};
-    std::vector<std::shared_ptr<DelegatedTargetTreeNode>> children;
-  };
-
-  class DelegationIterator {
-    using iterator_category = std::input_iterator_tag;
-    using value_type = Uptane::Target;
-    using difference_type = int;
-    using pointer = Uptane::Target *;
-    using reference = Uptane::Target &;
-
-   public:
-    explicit DelegationIterator(const Uptane::ImagesRepository &repo, std::shared_ptr<INvStorage> storage,
-                                std::shared_ptr<Uptane::Fetcher> fetcher, bool is_end = false);
-    DelegationIterator operator++();
-    bool operator==(const DelegationIterator &other) const;
-    bool operator!=(const DelegationIterator &other) const { return !(*this == other); }
-    const Uptane::Target &operator*();
-
-   private:
-    std::shared_ptr<DelegatedTargetTreeNode> tree_;
-    DelegatedTargetTreeNode *tree_node_;
-    const Uptane::ImagesRepository &repo_;
-    std::shared_ptr<INvStorage> storage_;
-    std::shared_ptr<Uptane::Fetcher> fetcher_;
-    std::shared_ptr<const Uptane::Targets> cur_targets_;
-    std::vector<Uptane::Targets>::size_type target_idx_{0};
-    std::vector<std::shared_ptr<DelegatedTargetTreeNode>>::size_type children_idx_{0};
-    bool terminating_{false};
-    int level_{0};
-    bool is_end_;
-
-    void renewTargetsData();
-  };
-
-  explicit LazyTargetsList(const Uptane::ImagesRepository &repo, std::shared_ptr<INvStorage> storage,
-                           std::shared_ptr<Uptane::Fetcher> fetcher)
-      : repo_{repo}, storage_{std::move(storage)}, fetcher_{std::move(fetcher)} {}
-  DelegationIterator begin() { return DelegationIterator(repo_, storage_, fetcher_); }
-  DelegationIterator end() { return DelegationIterator(repo_, storage_, fetcher_, true); }
-
- private:
-  const Uptane::ImagesRepository &repo_;
-  std::shared_ptr<INvStorage> storage_;
-  std::shared_ptr<Uptane::Fetcher> fetcher_;
-};
 
 class SotaUptaneClient {
  public:
@@ -109,11 +55,8 @@ class SotaUptaneClient {
   bool hasPendingUpdates();
   bool isInstallCompletionRequired();
   void completeInstall();
-  LazyTargetsList allTargets();
+  Uptane::LazyTargetsList allTargets();
 
-  static Uptane::Targets getTrustedDelegation(const Uptane::Role &delegate_role, const Uptane::Targets &parent_targets,
-                                              const Uptane::ImagesRepository &images_repo, INvStorage &storage,
-                                              Uptane::Fetcher &fetcher);
   bool updateImagesMeta();  // TODO: make private once aktualizr has a proper TUF API
   data::InstallationResult PackageInstall(const Uptane::Target &target);
